feat: dsn search bar + refactor
