+++ src/components/codeKeywords/index.ts
@@ -1,0 +1,5 @@
+export {
+  makeKeywordsClickable,
+  ORG_AUTH_TOKEN_REGEX,
+  KEYWORDS_REGEX,
+} from './codeKeywords';
